ml-service/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.0
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
numpy==1.24.3
requests==2.31.0
tiktoken
more-itertools
